second commit

+++ pa7/List.h
@@ -27,9 +27,9 @@ void freeList(List* pL); // Frees all heap memory associated with *pL, and sets
 // Access functions -----------------------------------------------------------
 int length(List L); // Returns the number of elements in L.
 int index(List L); // Returns index of cursor element if defined, -1 otherwise.
-int front(List L); // Returns front element of L. Pre: length()>0
-int back(List L); // Returns back element of L. Pre: length()>0
-int get(List L); // Returns cursor element of L. Pre: length()>0, index()>=0
+long front(List L); // Returns front element of L. Pre: length()>0
+long back(List L); // Returns back element of L. Pre: length()>0
+long get(List L); // Returns cursor element of L. Pre: length()>0, index()>=0
 int list_equals(List A, List B); // Returns true (1) iff Lists A and B are in same
 // state, and returns false (0) otherwise.
